update to use new RN 48 APIs. No longer works on older versions of RN
