+++ src/redux.js
@@ -1,0 +1,34 @@
+/* @flow */
+const NETWORK_CHANGED = 'redux-connectivity/NETWORK_CHANGED';
+
+type ActionType = {
+  type: string,
+  connected?: boolean,
+  connectionType?: string,
+};
+
+export function setNetworkAction(connected: boolean, type: string, effectiveType:string) {
+  return {
+    type: NETWORK_CHANGED,
+    connected,
+    type,
+    effectiveType,
+  };
+}
+
+const initialState = {
+};
+
+export default function reducer(state = initialState, action) {
+  switch (action.type) {
+    case NETWORK_CHANGED:
+      return {
+        ...state,
+        connected: action.connected,
+        type: action.type,
+        effectiveType: action.effectiveType,
+      };
+    default:
+      return state;
+  }
+}
